move crypto tests into separate files
