+++ soroban-sdk/src/tests.rs
@@ -16,7 +16,8 @@ mod contract_udt_struct;
 mod contract_udt_struct_tuple;
 mod contractimport;
 mod contractimport_with_error;
-mod crypto;
+mod crypto_ed25519;
+mod crypto_sha256;
 mod env;
 mod prng;
 mod proptest_scval_cmp;
